renamed test constants files
